+++ index.js
@@ -20,7 +20,7 @@ const { Stream } = require('./src/stream');
 const { Retry } = require('./src/retry');
 const { runPlan } = require('./src/run-plan');
 const { CircuitBreaker } = require('./src/circuit-breaker');
-const { wireGate, defaultActionTranslator } = require('./src/bareguard-adapter');
+const { wireGate, defaultActionTranslator, judgeToAnnotation } = require('./src/bareguard-adapter');
 const { toUnits, fromUnits, unitAssembler, unitTrimmer, harvestKey } = require('./src/context-units');
 const {
   BareAgentError,
@@ -64,6 +64,7 @@ module.exports = {
   CircuitBreaker,
   wireGate,
   defaultActionTranslator,
+  judgeToAnnotation,
   toUnits,
   fromUnits,
   unitAssembler,